Fixed minor bug related with type conversions

+++ Flutter/whatz_up/lib/models/event.dart
@@ -17,10 +17,10 @@ class Event {
     required this.longitude,
   });
 
-
   static FirebaseFunctions functions = FirebaseFunctions.instance;
 
-  static Future<Event> create(String name, String description, double price, double latitude, double longitude) async {
+  static Future<Event> create(String name, String description, double price,
+      double latitude, double longitude) async {
     Map<String, dynamic> newEvent = {
       'name': name,
       'description': description,
@@ -28,7 +28,8 @@ class Event {
       'latitude': latitude,
       'longitude': longitude
     };
-    final response = (await functions.httpsCallable('create_event').call(newEvent)).data;
+    final response =
+        (await functions.httpsCallable('create_event').call(newEvent)).data;
 
     return Event(
       id: response['id'],
@@ -41,7 +42,8 @@ class Event {
   }
 
   static Future<Event> fromId(String id) async {
-    final response = (await functions.httpsCallable('get_event').call({ 'id': id })).data;
+    final response =
+        (await functions.httpsCallable('get_event').call({'id': id})).data;
 
     return Event(
       id: response['id'],
@@ -54,15 +56,21 @@ class Event {
   }
 
   static Future<List<Event>> find() async {
-    final List<dynamic> response = (await functions.httpsCallable('get_events').call()).data;
+    final List<dynamic> response =
+        (await functions.httpsCallable('get_events').call()).data;
 
-    return response.map((event) => Event(
-      id: event['id'],
-      name: event['name'],
-      description: event['description'],
-      price: event['price'],
-      latitude: event['latitude'],
-      longitude: event['longitude'],
-    )).toList();
+    return response
+        .map((event) => Event(
+              id: event['id'],
+              name: event['name'],
+              description: event['description'],
+              price: (event['price'] as int)
+                  .toDouble(), // TODO: Temporary fix, should be fixed in the database
+              latitude: (event['latitude'] as int)
+                  .toDouble(), // TODO: Temporary fix, should be fixed in the database
+              longitude: (event['longitude'] as int)
+                  .toDouble(), // TODO: Temporary fix, should be fixed in the database
+            ))
+        .toList();
   }
 }
